Fixed Albin's remark: In order to generate a BOM from KiCAD using InteractiveHtmlBom, all components which should not be mounted must be tagged in the netlist or these will also show up in the BOM.
The name of the tag could be anything, my proposal is DNP (Do Not Populate). The tag is chosen in a drop-down list when generating the BOM. All components which has a non-empty tag of this name will be ignored/blacklisted.

This is done by inserting a field name and value in the netlist. Example:

(comp (ref A-REF101)
  (value HMM013/2)
  (footprint HM-REFERENCE-POINT-0.8)
  (fields
    (field (name DNP) "##"))
)